fix: 업체 로그아웃 로직 및 레이아웃 수정
