+++ app/(main)/customer/my-page/[id]/_components/MyPageContents/MyPageContents.tsx
@@ -14,8 +14,7 @@ import KakaoMap from '@/app/_components/KakaoMap/KakaoMap';
 import { profileType } from '../../../../../../_types/member/profileType';
 import { patchMember } from '@/app/_services/member/patchMember';
 import { getMember } from '@/app/_services/member/getMember';
-import { useAuth } from '@/app/_providers/AuthProvider';
-import PopUp from '@/app/_components/pop-up/PopUp';
+import LogoutButton from '@/app/_components/logout-button/LogoutButton';
 
 type initialValuesType = {
   nickName: string;
@@ -39,10 +38,8 @@ const MyPageContents = () => {
   const [address, setAddress] = useState(profile.address.name);
   const [addressError, setAddressError] = useState(true);
   const [click, setClick] = useState(false);
-  const [open, setOpen] = useState(false);
 
   const coords = useCoordinate(address);
-  const { logout } = useAuth();
 
   const schema = object().shape({
     nickName: isValidNickName(),
@@ -169,27 +166,7 @@ const MyPageContents = () => {
       <PrimaryButton className="mb-5" type="submit" onClick={changeProfile}>
         정보 수정하기
       </PrimaryButton>
-      <button
-        style={{ boxShadow: '0px 0px 4px 0px rgb(0, 0, 0, 0.1)' }}
-        className="mb-5 h-12 w-full rounded-md bg-light-gray text-base text-red"
-        onClick={() => {
-          setOpen(true);
-        }}
-      >
-        로그아웃
-      </button>
-      {open && (
-        <PopUp
-          mainText="로그아웃하시겠습니까?"
-          leftBtnText="예"
-          leftBtnClick={() => {
-            setOpen(false);
-            logout();
-          }}
-          rightBtnText="아니요"
-          rightBtnClick={() => setOpen(false)}
-        />
-      )}
+      <LogoutButton />
     </form>
   );
 };
